requirements.txt
Django==1.10.5
Pillow==3.4.2
beautifulsoup4==4.5.1
django-watson==1.3.0
django-anymail==0.6.1
django-crontab==0.7.1
django-admin-view-permission==0.9
djangorestframework==3.6.3